Moved Evaluable Expression Fitness Function.
